requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pinecone-client==2.2.1
langchain==0.0.308
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2
tenacity==8.2.3
PyPDF2==3.0.1
requests==2.31.0
gunicorn